converted badge emphasis to enum
