+++ src/components/Badge/index.ts
@@ -1,1 +1,2 @@
 export * from "./Badge";
+export * from "./types";
